requirements-advanced.txt
# Enhanced Requirements for Advanced Deriv SyntX Bot

# Core Dependencies
python-telegram-bot>=20.0
python-deriv-api>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# AI/ML Dependencies
torch>=2.0.0
torchvision>=0.15.0

# Technical Analysis (optional - use with caution on Python 3.14)
# pandas-ta>=0.3.14b0

# Database
sqlite3  # Built-in with Python

# Async Support
asyncio  # Built-in with Python

# Logging
logging  # Built-in with Python
